refactor(components): refactor divider
